add option for disabling arrow icon in collapse panel
