+++ components/collapse/demo/noarrow.md
@@ -1,0 +1,40 @@
+---
+order: 4
+title:
+  zh-CN: TODO
+  en-US: No arrow
+---
+
+## zh-CN
+
+TODO
+
+## en-US
+
+You can disable showing arrow icon by passing `showArrow={false}` to `CollapsePanel` component.
+
+````jsx
+import { Collapse } from 'antd';
+const Panel = Collapse.Panel;
+
+function callback(key) {
+  console.log(key);
+}
+
+const text = `
+  A dog is a type of domesticated animal.
+  Known for its loyalty and faithfulness,
+  it can be found as a welcome guest in many households across the world.
+`;
+
+ReactDOM.render(
+  <Collapse defaultActiveKey={['1']} onChange={callback}>
+    <Panel header="This is panel header with arrow icon" key="1">
+      <p>{text}</p>
+    </Panel>
+    <Panel showArrow={false} header="This is panel header with no arrow icon" key="2">
+      <p>{text}</p>
+    </Panel>
+  </Collapse>
+, mountNode);
+````
