+++ components/collapse/index.en-US.md
@@ -29,3 +29,4 @@ A content area which can be collapsed and expanded.
 | disabled | If `true`, panel cannot be opened or closed | boolean | `false` |
 | header | Title of the panel | string\|ReactNode | - |
 | key | Unique key identifying the panel from among its siblings | string | - |
+| showArrow | If `false`, panel will not show arrow icon | boolean | `true` |
