+++ components/collapse/index.tsx
@@ -1,5 +1,6 @@
 import Collapse from './Collapse';
 
-export { CollapsePanelProps, CollapseProps } from './Collapse';
+export { CollapseProps } from './Collapse';
+export { CollapsePanelProps } from './CollapsePanel';
 
 export default Collapse;
