+++ components/collapse/style/index.less
@@ -54,6 +54,12 @@
         }
       }
     }
+
+    &.@{collapse-prefix-cls}-no-arrow {
+      > .@{collapse-prefix-cls}-header {
+        padding-left: 12px;
+      }
+    }
   }
 
   &-anim-active {
